api/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
openpyxl
fpdf2
python-multipart
